+++ src/main/java/org/bot/nullbot/dispatcher/listener/CommandListener.java
@@ -36,14 +36,14 @@ public class CommandListener
     {
         // 串行调用 消息预处理
         monitorListener.onGroupKeywordDetection(bot, event);
-        if(!monitorListener.onGroupAIAutoReply(bot, event))  // 触发自动发言会记录当前消息 忽略消息收集
+        if (!monitorListener.onGroupAIAutoReply(bot, event))  // 触发自动发言会记录当前消息 忽略消息收集
             monitorListener.onGroupMessageCollection(bot, event);
         monitorListener.onGroupImageCollection(bot, event);
 
         if (event.getMessage().startsWith(commandPrefix)) {  // 检测普通命令
             log.info("◉ [GroupAction:Command] 来自群 {} - {}({}) -> {}", event.getGroupId(), event.getSender().getNickname(), event.getSender().getUserId(), event.getMessage().replaceAll("\\R", " "));
             commandProcessor.processQQ(bot, new CommandEvent<>(event));
-        }else if(event.getArrayMsg().size() >= 2 && event.getArrayMsg().get(0).getType() == MsgTypeEnum.reply){  // 检测引用命令
+        } else if (event.getArrayMsg().size() >= 2 && event.getArrayMsg().get(0).getType() == MsgTypeEnum.reply){  // 检测引用命令
             String slashCommand = event.getArrayMsg().get(1).getData().get("text");
             if(slashCommand != null && slashCommand.startsWith(commandPrefix)){
                 log.info("◉ [GroupAction:ReplyCommand] 来自群 {} - {}({}) -> {}", event.getGroupId(), event.getSender().getNickname(), event.getSender().getUserId(), event.getMessage().replaceAll("\\R", " "));
@@ -59,7 +59,7 @@ public class CommandListener
     {
         // 串行调用 消息预处理
         // monitorListener.onGroupKeywordDetection(bot, event);  // 禁用 关键词检测
-        // if(!monitorListener.onGroupAIAutoReply(bot, event))  // 无需调用 AI即将回复
+        // if (!monitorListener.onGroupAIAutoReply(bot, event))  // 无需调用 AI即将回复
         //     monitorListener.onGroupMessageCollection(bot, event);  // 无需调用 AI自动记录
         monitorListener.onGroupImageCollection(bot, event);